+++ src/GSS.Authorization.OAuth2.HttpClient/OAuth2HttpHandler.cs
@@ -1,10 +1,13 @@
 using System;
+using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Caching.Memory;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Logging.Abstractions;
 
 namespace GSS.Authorization.OAuth2
 {
@@ -13,13 +16,20 @@ namespace GSS.Authorization.OAuth2
         private readonly SemaphoreSlim _semaphore = new SemaphoreSlim(1, 1);
         private readonly IAuthorizer _authorizer;
         private readonly IMemoryCache _memoryCache;
+        private readonly ILogger<OAuth2HttpHandler> _logger;
         private readonly string _cacheKey;
 
-        public OAuth2HttpHandler(IAuthorizer authorizer, IMemoryCache memoryCache)
+        public OAuth2HttpHandler(IAuthorizer authorizer, IMemoryCache memoryCache, ILoggerFactory loggerFactory)
         {
             _authorizer = authorizer;
             _memoryCache = memoryCache;
+            _logger = loggerFactory.CreateLogger<OAuth2HttpHandler>();
             _cacheKey = Guid.NewGuid().ToString();
+        }
+
+        [Obsolete("Use another constructor")]
+        public OAuth2HttpHandler(IAuthorizer authorizer, IMemoryCache memoryCache) : this(authorizer, memoryCache, NullLoggerFactory.Instance)
+        {
         }
 
         protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request,
@@ -32,8 +42,11 @@ namespace GSS.Authorization.OAuth2
                 TrySetAuthorizationHeaderToRequest(await GetAccessTokenAsync(cancellationToken).ConfigureAwait(false), request);
             }
             var response = await base.SendAsync(request, cancellationToken).ConfigureAwait(false);
-            if (response.StatusCode != HttpStatusCode.Unauthorized)
+            // https://tools.ietf.org/html/rfc6750#section-3
+            var authenticateError = response.Headers.WwwAuthenticate.FirstOrDefault()?.Parameter;
+            if (string.IsNullOrWhiteSpace(authenticateError) && response.StatusCode != HttpStatusCode.Unauthorized)
                 return response;
+            _logger.LogError(authenticateError);
             TrySetAuthorizationHeaderToRequest(await GetAccessTokenAsync(cancellationToken, forceRenew: true).ConfigureAwait(false), request);
             return await base.SendAsync(request, cancellationToken).ConfigureAwait(false);
         }
@@ -70,10 +83,10 @@ namespace GSS.Authorization.OAuth2
 
         private static void TrySetAuthorizationHeaderToRequest(AccessToken accessToken, HttpRequestMessage request)
         {
-            if (accessToken != null && !string.IsNullOrWhiteSpace(accessToken.Token))
+            if (!string.IsNullOrWhiteSpace(accessToken?.Token))
             {
                 request.Headers.Authorization =
-                    new AuthenticationHeaderValue(AuthorizerDefaults.Bearer, accessToken.Token);
+                    new AuthenticationHeaderValue(AuthorizerDefaults.Bearer, accessToken?.Token);
             }
         }
     }